+++ app/Notifications/VerifyEmail.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace App\Notifications;
+
+use Carbon\Carbon;
+use Illuminate\Support\Facades\URL;
+use Illuminate\Auth\Notifications\VerifyEmail as Notification;
+
+class VerifyEmail extends Notification
+{
+    /**
+     * Get the verification URL for the given notifiable.
+     *
+     * @param  mixed  $notifiable
+     * @return string
+     */
+    protected function verificationUrl($notifiable)
+    {
+        $appUrl = config('app.client_url', config('app.url'));
+
+        $url = URL::temporarySignedRoute(
+            'verification.verify', Carbon::now()->addMinutes(60), ['user' => $notifiable->id]
+        );
+
+        return str_replace(url('/api'), $appUrl, $url);
+    }
+}