+++ client/components/global/index.js
@@ -2,7 +2,7 @@ import Vue from 'vue'
 
 const requireContext = require.context('./', false, /.*\.(vue)$/)
 
-requireContext.keys().forEach(file => {
+requireContext.keys().forEach((file) => {
   const Component = requireContext(file).default
 
   if (Component.name) {