+++ client/layouts/simple.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="simple-layout d-flex align-items-center justify-content-center m-0 bg-white">
-    <nuxt/>
+    <nuxt />
   </div>
 </template>