+++ client/pages/auth/register.vue
@@ -1,15 +1,19 @@
 <template>
   <div class="row">
     <div class="col-lg-8 m-auto">
-      <card :title="$t('register')">
+      <card v-if="mustVerifyEmail" :title="$t('register')">
+        <div class="alert alert-success" role="alert">
+          {{ $t('verify_email_address') }}
+        </div>
+      </card>
+      <card v-else :title="$t('register')">
         <form @submit.prevent="register" @keydown="form.onKeydown($event)">
           <!-- Name -->
           <div class="form-group row">
             <label class="col-md-3 col-form-label text-md-right">{{ $t('name') }}</label>
             <div class="col-md-7">
-              <input v-model="form.name" :class="{ 'is-invalid': form.errors.has('name') }" type="text" name="name"
-                     class="form-control">
-              <has-error :form="form" field="name"/>
+              <input v-model="form.name" :class="{ 'is-invalid': form.errors.has('name') }" type="text" name="name" class="form-control">
+              <has-error :form="form" field="name" />
             </div>
           </div>
 
@@ -17,9 +21,8 @@
           <div class="form-group row">
             <label class="col-md-3 col-form-label text-md-right">{{ $t('email') }}</label>
             <div class="col-md-7">
-              <input v-model="form.email" :class="{ 'is-invalid': form.errors.has('email') }" type="email" name="email"
-                     class="form-control">
-              <has-error :form="form" field="email"/>
+              <input v-model="form.email" :class="{ 'is-invalid': form.errors.has('email') }" type="email" name="email" class="form-control">
+              <has-error :form="form" field="email" />
             </div>
           </div>
 
@@ -27,9 +30,8 @@
           <div class="form-group row">
             <label class="col-md-3 col-form-label text-md-right">{{ $t('password') }}</label>
             <div class="col-md-7">
-              <input v-model="form.password" :class="{ 'is-invalid': form.errors.has('password') }" type="password" name="password"
-                     class="form-control">
-              <has-error :form="form" field="password"/>
+              <input v-model="form.password" :class="{ 'is-invalid': form.errors.has('password') }" type="password" name="password" class="form-control">
+              <has-error :form="form" field="password" />
             </div>
           </div>
 
@@ -38,8 +40,9 @@
             <label class="col-md-3 col-form-label text-md-right">{{ $t('confirm_password') }}</label>
             <div class="col-md-7">
               <input v-model="form.password_confirmation" :class="{ 'is-invalid': form.errors.has('password_confirmation') }" type="password" name="password_confirmation"
-                     class="form-control">
-              <has-error :form="form" field="password_confirmation"/>
+                     class="form-control"
+              >
+              <has-error :form="form" field="password_confirmation" />
             </div>
           </div>
 
@@ -51,7 +54,7 @@
               </v-button>
 
               <!-- GitHub Login Button -->
-              <login-with-github/>
+              <login-with-github />
             </div>
           </div>
         </form>
@@ -74,7 +77,8 @@ export default {
       email: '',
       password: '',
       password_confirmation: ''
-    })
+    }),
+    mustVerifyEmail: false
   }),
 
   methods: {
@@ -82,17 +86,22 @@ export default {
       // Register the user.
       const { data } = await this.form.post('/register')
 
-      // Log in the user.
-      const { data: { token } } = await this.form.post('/login')
+      // Must verify email fist.
+      if (data.status) {
+        this.mustVerifyEmail = true
+      } else {
+        // Log in the user.
+        const { data: { token } } = await this.form.post('/login')
 
-      // Save the token.
-      this.$store.dispatch('auth/saveToken', { token })
+        // Save the token.
+        this.$store.dispatch('auth/saveToken', { token })
 
-      // Update the user.
-      await this.$store.dispatch('auth/updateUser', { user: data })
+        // Update the user.
+        await this.$store.dispatch('auth/updateUser', { user: data })
 
-      // Redirect home.
-      this.$router.push({ name: 'home' })
+        // Redirect home.
+        this.$router.push({ name: 'home' })
+      }
     }
   }
 }