+++ client/router.js
@@ -4,33 +4,25 @@ import { scrollBehavior } from '~/utils'
 
 Vue.use(Router)
 
-const Home = () => import('~/pages/home').then(m => m.default || m)
-const Welcome = () => import('~/pages/welcome').then(m => m.default || m)
-
-const Login = () => import('~/pages/auth/login').then(m => m.default || m)
-const Register = () => import('~/pages/auth/register').then(m => m.default || m)
-const PasswordReset = () => import('~/pages/auth/password/reset').then(m => m.default || m)
-const PasswordRequest = () => import('~/pages/auth/password/email').then(m => m.default || m)
-
-const Settings = () => import('~/pages/settings/index').then(m => m.default || m)
-const SettingsProfile = () => import('~/pages/settings/profile').then(m => m.default || m)
-const SettingsPassword = () => import('~/pages/settings/password').then(m => m.default || m)
+const page = path => () => import(`~/pages/${path}`).then(m => m.default || m)
 
 const routes = [
-  { path: '/', name: 'welcome', component: Welcome },
-  { path: '/home', name: 'home', component: Home },
+  { path: '/', name: 'welcome', component: page('welcome.vue') },
 
-  { path: '/login', name: 'login', component: Login },
-  { path: '/register', name: 'register', component: Register },
-  { path: '/password/reset', name: 'password.request', component: PasswordRequest },
-  { path: '/password/reset/:token', name: 'password.reset', component: PasswordReset },
+  { path: '/login', name: 'login', component: page('auth/login.vue') },
+  { path: '/register', name: 'register', component: page('auth/register.vue') },
+  { path: '/password/reset', name: 'password.request', component: page('auth/password/email.vue') },
+  { path: '/password/reset/:token', name: 'password.reset', component: page('auth/password/reset.vue') },
+  { path: '/email/verify/:id', name: 'verification.verify', component: page('auth/verification/verify.vue') },
+  { path: '/email/resend', name: 'verification.resend', component: page('auth/verification/resend.vue') },
 
+  { path: '/home', name: 'home', component: page('home.vue') },
   { path: '/settings',
-    component: Settings,
+    component: page('settings/index.vue'),
     children: [
       { path: '', redirect: { name: 'settings.profile' } },
-      { path: 'profile', name: 'settings.profile', component: SettingsProfile },
-      { path: 'password', name: 'settings.password', component: SettingsPassword }
+      { path: 'profile', name: 'settings.profile', component: page('settings/profile.vue') },
+      { path: 'password', name: 'settings.password', component: page('settings/password.vue') }
     ] }
 ]