+++ tests/Feature/LocaleTest.php
@@ -10,7 +10,7 @@ class LocaleTest extends TestCase
     public function set_locale_from_header()
     {
         $this->withHeaders(['Accept-Language' => 'zh-CN'])
-            ->postJson('/login');
+            ->postJson('/api/login');
 
         $this->assertEquals('zh-CN', $this->app->getLocale());
     }
@@ -19,7 +19,7 @@ class LocaleTest extends TestCase
     public function set_locale_from_header_short()
     {
         $this->withHeaders(['Accept-Language' => 'en-US'])
-            ->postJson('/login');
+            ->postJson('/api/login');
 
         $this->assertEquals('en', $this->app->getLocale());
     }