+++ tests/Feature/RegisterTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Feature;
 
+use App\User;
 use Tests\TestCase;
 
 class RegisterTest extends TestCase
@@ -9,7 +10,7 @@ class RegisterTest extends TestCase
     /** @test */
     public function can_register()
     {
-        $this->postJson('/register', [
+        $this->postJson('/api/register', [
             'name' => 'Test User',
             'email' => 'test@test.app',
             'password' => 'secret',
@@ -17,5 +18,20 @@ class RegisterTest extends TestCase
         ])
         ->assertSuccessful()
         ->assertJsonStructure(['id', 'name', 'email']);
+    }
+
+    /** @test */
+    public function can_not_register_with_existing_email()
+    {
+        factory(User::class)->create(['email' => 'test@test.app']);
+
+        $this->postJson('/api/register', [
+            'name' => 'Test User',
+            'email' => 'test@test.app',
+            'password' => 'secret',
+            'password_confirmation' => 'secret',
+        ])
+        ->assertStatus(422)
+        ->assertJsonValidationErrors(['email']);
     }
 }